feat: Update models and routes for SMTP configuration, application general settings, user dashboard settings, dashboard network settings, network speed test result, and feature toggle settings
