+++ src/models/smtp_configuration.py
@@ -1,0 +1,8 @@
+from src.config import db
+
+
+class SMTPSettings(db.Model):
+    __tablename__ = "smtp_settings"
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(150), unique=True, nullable=False)
+    password = db.Column(db.String(150), nullable=False)
