+++ src/routes/disk_info.py
@@ -4,14 +4,14 @@ from flask import render_template, blueprints, flash
 from flask_login import login_required
 from src.config import app
 from src.utils import get_cached_value, get_disk_free, get_disk_total, get_disk_used, get_disk_usage_percent
-from src.models import DashboardSettings, FeatureTogglesSettings
+from src.models import FeatureToggleSettings
 
 disk_info_bp = blueprints.Blueprint("disk_usage", __name__)
 
 @app.route("/disk_usage")
 @login_required
 def disk_usage():
-    feature_toggles_settings = FeatureTogglesSettings.query.first()
+    feature_toggles_settings = FeatureToggleSettings.query.first()
     if not feature_toggles_settings.is_disk_info_enabled:
         flash("You do not have permission to view this page.", "danger")
         return render_template("error/permission_denied.html")
