+++ src/routes/memory_info.py
@@ -4,7 +4,7 @@ from flask_login import login_required
 
 from src.config import app
 from src.utils import get_cached_value, get_memory_percent, get_memory_available, get_memory_used, get_swap_memory_info
-from src.models import DashboardSettings, FeatureTogglesSettings
+from src.models import FeatureToggleSettings
 
 memory_info_bp = blueprints.Blueprint("memory_usage", __name__)
 
@@ -12,7 +12,7 @@ memory_info_bp = blueprints.Blueprint("memory_usage", __name__)
 @app.route("/memory_usage")
 @login_required
 def memory_usage():
-    feature_toggles_settings = FeatureTogglesSettings.query.first()
+    feature_toggles_settings = FeatureToggleSettings.query.first()
     if not feature_toggles_settings.is_memory_info_enabled:
         flash("You do not have permission to view this page.", "danger")
         return render_template("error/permission_denied.html")
