+++ src/routes/network_info.py
@@ -3,14 +3,14 @@ from flask import render_template, blueprints, flash
 from flask_login import login_required
 from src.config import app
 from src.utils import get_established_connections
-from src.models import DashboardSettings, FeatureTogglesSettings
+from src.models import  FeatureToggleSettings
 
 network_info_bp = blueprints.Blueprint("network_stats", __name__)
 
 @app.route("/network_stats")
 @login_required
 def network_stats():
-    feature_toggles_settings = FeatureTogglesSettings.query.first()
+    feature_toggles_settings = FeatureToggleSettings.query.first()
     if not feature_toggles_settings.is_network_info_enabled:
         flash("You do not have permission to view this page.", "danger")
         return render_template("error/permission_denied.html")
