+++ src/routes/process.py
@@ -3,7 +3,7 @@ from flask import request, render_template, redirect, url_for, flash, session, b
 from flask_login import login_required, current_user
 from src.config import app
 from src.utils import get_top_processes, render_template_from_file
-from src.models import DashboardSettings, FeatureTogglesSettings
+from src.models import  FeatureToggleSettings
 from src.scripts.email_me import send_smpt_email
 
 process_bp = blueprints.Blueprint("process", __name__)
@@ -11,7 +11,7 @@ process_bp = blueprints.Blueprint("process", __name__)
 @app.route("/process", methods=["GET", "POST"])
 @login_required
 def process():
-    feature_toggles_settings = FeatureTogglesSettings.query.first()
+    feature_toggles_settings = FeatureToggleSettings.query.first()
     if not feature_toggles_settings.is_process_info_enabled:
         flash("You do not have permission to view this page.", "danger")
         return render_template("error/permission_denied.html")
